+++ hydromt/utils.py
@@ -3,6 +3,11 @@ import os
 import numpy as np
 
 
+class _classproperty(property):
+    def __get__(self, owner_self, owner_cls):
+        return self.fget(owner_cls)
+
+
 def partition_dictionaries(left, right):
     """Calculate a partitioning of the two dictionaries.